refactor: Меняет источник получение моков фильмов на redux, выносит currentMoviePage в отдельный компонент
