+++ src/components/app/app.jsx
@@ -1,17 +1,17 @@
 import React from "react";
 import PropTypes from "prop-types";
 import {BrowserRouter, Switch, Route} from "react-router-dom";
+import {connect} from "react-redux";
 
 import MainPage from "../main-page/main-page";
 import SignInPage from "../sign-in-page/sign-in-page";
 import MylistPage from "../mylist-page/mylist-page";
-import MoviePage from "../movie-page/movie-page";
 import ReviewPage from "../review-page/review-page";
 import PlayerPage from "../player-page/player-page";
+import CurrentMoviePage from "../currentMoviePage/currentMoviePage";
 
 import moviesProp from "../movie-card/movie-card.prop";
-
-const RELATED_MOVIES_COUNT = 4;
+import reviewProp from "../review/review.prop";
 
 const App = ({movies, reviews}) => {
   return (
@@ -19,9 +19,7 @@ const App = ({movies, reviews}) => {
       <Switch>
 
         <Route path="/" exact>
-          <MainPage
-            movies={movies}
-          />
+          <MainPage movies={movies} />
         </Route>
 
         <Route path="/sign-in" exact>
@@ -29,25 +27,19 @@ const App = ({movies, reviews}) => {
         </Route>
 
         <Route path="/mylist" exact>
-          <MylistPage
-            movies={movies}
-          />
+          <MylistPage movies={movies} />
         </Route>
 
         <Route
           path="/films/:id"
           exact
-          render={({match: {params: {id}}}) => {
-            const [currentMovie] = movies.filter((movie) => movie.id === id);
-            const relatedMovies = movies.filter((movie) => movie.genre === currentMovie.genre);
-            const movieInfo = Object.assign({}, currentMovie, {reviews});
-            return (
-              <MoviePage
-                movieInfo={movieInfo}
-                related={relatedMovies.slice(0, RELATED_MOVIES_COUNT)}
-              />);
-          }
-          }
+          render={
+            ({match: {params: {id}}}) =>
+              <CurrentMoviePage
+                pageId={id}
+                movies={movies}
+                reviews={reviews}
+              />}
         />
 
         <Route path="/films/:id/review" exact>
@@ -63,9 +55,13 @@ const App = ({movies, reviews}) => {
   );
 };
 
+const mapStateToProps = (state) => ({
+  movies: state.movies
+});
+
 App.propTypes = {
   movies: PropTypes.arrayOf(PropTypes.shape(moviesProp)).isRequired,
-  reviews: PropTypes.array.isRequired,
+  reviews: PropTypes.arrayOf(PropTypes.shape(reviewProp)),
 };
 
-export default App;
+export default connect(mapStateToProps)(App);
